fix delete tenant user
If you tried to delete the default tenantUser and there was no other entity available to mark as default, the previous behavior was to delete the user and then delete the tenantUser entity.

Since the deletion method is configured in a cascade form, this logic could result in trying to delete the tenantUser and this entity would no longer exist, leading to errors.

I changed this behavior to first delete the tenantUser entity and then the user.

In addition, a description of this process was added to the documentation and a unit test was added that, among other cases, also tests this possibility.